ecommerce/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
Markdown==3.6
PyJWT==2.8.0
requests==2.31.0
sqlparse==0.5.0
urllib3==2.2.1
# psycopg2==2.9.9
